API

Program

Pipelines.Program — Type

Summary

abstract type Program <: Any

Subtypes

CmdProgram
JuliaProgram

Command Program

Pipelines.CmdProgram — Method

CmdProgram <: Program

CmdProgram(;
	name::String               = "Command Program",
	id_file::String            = "",
	info_before::String        = "auto",
	info_after::String         = "auto",
	cmd_dependencies::Vector{CmdDependency} = Vector{CmdDependency}(),
	inputs                     = Vector{String}(),
	validate_inputs::Function  = do_nothing,  # positional arguments: inputs::Dict{String, ValidInputTypes}
	prerequisites::Function    = do_nothing,  # positional arguments: inputs, outputs::Dict{String, ValidInputTypes}
	cmd::Base.AbstractCmd      = ``,
	infer_outputs::Function    = do_nothing,  # positional arguments: inputs::Dict{String, ValidInputTypes}
	outputs                    = Vector{String}(),
	validate_outputs::Function = do_nothing  # positional arguments: outputs::Dict{String, ValidInputTypes},
	wrap_up::Function          = do_nothing  # positional arguments: inputs, outputs::Dict{String, ValidInputTypes}
) -> CmdProgram

Command program template. To run a CmdProgram, use run(::CmdProgram; kwargs...).

Arguments

name::String: Program name.

id_file::String: The prefix of run ID file. To prevent from running the program with the same inputs and outputs twice, it will generate a unique run ID file after a successful run.

info_before::String: Print it when the program is started.

info_after::String: Print it when the program is finished.

cmd_dependencies::Vector{CmdDependency}: Any command dependencies used in the program.

inputs and outputs: Elements (or vectors containing elements) in the following format: (1) keyword (2) keyword => data_type (3) keyword => default_value (4) keyword => default_value => data_type.

keyword is an argument name, can be String or Symbol.

default_value is optional. If set, users may not provide this argument when running. Elsewise, users have to provide it. Caution: nothing is preserved and means default value not set. If String, it can contain other keywords, but need to quote using '<>', such as "<arg>.txt"

data_type is optional. If set, the value provided have to be this data type, or an error will throw.

HOW DOES THIS WORK?

CmdProgram stores a command template. In the template, replaceable portions are occupied by keywords, and all keywords are set in inputs and outputs. keywords will be replaced before running the program. Users need to provide a dictionary of keyword::String => value in run(::Program, inputs::Dict{String}, outputs::Dict{String}).

validate_inputs::Function: A function to validate inputs. It takes one argument Dict{String, ValidInputTypes} whose keys are the same as inputs. If validation fail, throw error or return false.

prerequisites: A function to run just before the main command. It prepares necessary things, such as creating directories. It takes two arguments Dict{String, ValidInputTypes} whose keys are the same as inputs and outputs, respectively.

cmd::AbstractCmd: The main command template. In the template, keywords in inputs::Vector{String} and outputs::Vector{String} will be replaced when envoking run(::CmdProgram, inputs::Dict{String, ValidInputTypes}, outputs::Dict{String, ValidInputTypes}).

infer_outputs::Function: A function to infer outputs from inputs. It takes one argument Dict{String, ValidInputTypes} whose keys are the same as inputs.

validate_outputs::Function: A function to validate outputs. It takes one argument Dict{String, ValidInputTypes} whose keys are the same as outputs. If validation fail, throw error or return false.

wrap_up::Function: the last function to run. It takes two arguments Dict{String, ValidInputTypes} whose keys are the same as inputs and outputs, respectively.

Example

p = CmdProgram(
	id_file = "id_file",
	inputs = [
		"input",
		"input2" => Int,
		"optional_arg" => 5,
		"optional_arg2" => 0.5 => Number
	],
	outputs =
		"output" => "<input>.output"
	,
	cmd = `echo input input2 optional_arg optional_arg2 output`
)

inputs = Dict(
	"input" => `in1`,
	"input2" => 2
)

outputs = Dict(
	"output" => "out"
)

run(p, inputs, outputs;
	touch_run_id_file = false
)

Pipelines.prepare_cmd — Function

prepare_cmd(p::CmdProgram, inputs, outputs)

Prepare the runable command. Keywords in CmdProgram will be given to values of inputs/outputs.

Julia Program

Pipelines.JuliaProgram — Method

JuliaProgram <: Program

JuliaProgram(;
	name::String               = "Unnamed Command Program",
	id_file::String            = "",
	info_before::String        = "auto",
	info_after::String         = "auto",
	cmd_dependencies::Vector{CmdDependency} = Vector{CmdDependency}(),
	inputs                     = Vector{String}(),
	validate_inputs::Function  = do_nothing,  # positional arguments: inputs::Dict{String}
	prerequisites::Function    = do_nothing,  # positional arguments: inputs, outputs::Dict{String}
	main::Function             = do_nothing,  # positional arguments: inputs, outputs::Dict{String}
	infer_outputs::Function    = do_nothing,  # positional arguments: inputs::Dict{String}
	outputs                    = Vector{String}(),
	validate_outputs::Function = do_nothing,  # positional arguments: outputs::Dict{String}
	wrap_up::Function          = do_nothing  # positional arguments: inputs, outputs::Dict{String}
) -> JuliaProgram

Julia program template. To run a JuliaProgram, use run(::JuliaProgram; kwargs...).

Arguments

name::String: Program name.

id_file::String: The prefix of run ID file. To prevent from running the program with the same inputs and outputs twice, it will generate a unique run ID file after a successful run.

info_before::String: Print it when the program is started.

info_after::String: Print it when the program is finished.

cmd_dependencies::Vector{CmdDependency}: Any command dependencies used in the program.

inputs and outputs: Elements (or vectors containing elements) in the following format: (1) keyword (2) keyword => data_type (3) keyword => default_value (4) keyword => default_value => data_type.

keyword is an argument name, can be String or Symbol.

default_value is optional. If set, users may not provide this argument when running. Elsewise, users have to provide it. Caution: nothing is preserved and means default value not set. If String, it can contain other keywords, but need to quote using '<>', such as "<arg>.txt"

data_type is optional. If set, the value provided have to be this data type, or an error will throw.

HOW DOES THIS WORK?

JuliaProgram stores a Julia function, with two arguments inputs::Dict{String}, outputs::Dict{String}. The keys of the two arguments should be set in inputs::Vector{String} and outputs::Vector{String}. Caution: the returned value of p.main will be assigned to new outputs. Please ensure the returned value is Dict{String} with proper keys.

validate_inputs::Function: A function to validate inputs. It takes one argument Dict{String} whose keys are the same as inputs. If validation fail, throw error or return false.

prerequisites: A function to run just before the main command. It prepares necessary things, such as creating directories. It takes two arguments Dict{String} whose keys are the same as inputs and outputs, respectively.

main::Function: The main julia function. It takes two arguments Dict{String} whose keys are the same as inputs and outputs, respectively.

Caution: the returned value of p.main will be assigned to new outputs. Please ensure the returned value is Dict{String} with proper keys.

infer_outputs::Function: A function to infer outputs from inputs. It takes one argument Dict{String} whose keys are the same as inputs.

validate_outputs::Function: A function to validate outputs. It takes one argument Dict{String} whose keys are the same as outputs. If validation fail, throw error or return false.

wrap_up::Function: the last function to run. It takes two arguments Dict{String} whose keys are the same as inputs and outputs, respectively.

Example

p = JuliaProgram(
	id_file = "id_file",
	inputs = [
		"a",
		"b" => Int
	],
	outputs =
		"c" => "<a>.<b>"
	,
	main = (inputs, outputs) -> begin
		a = inputs["a"]
		b = inputs["b"]
		println("inputs are ", a, " and ", b)
		println("You can also use info in outputs: ", outputs["c"])
        println("The returned value will be assigned to a new outputs")

        return Dict{String,Any}("c" => b^2)
	end
)

inputs = Dict(
	"a" => `in1`,
	"b" => 2
)

outputs = Dict(
	"c" => "out"
)

success, outputs = run(p, inputs, outputs;
	touch_run_id_file = false
) # outputs will be refreshed

Run Programs

Base.run — Function

run(
	p::CmdProgram;
	inputs=Dict{String}(),
	outputs=Dict{String}(),
	dir::AbstractString="",
	check_dependencies::Bool=true,
	skip_when_done::Bool=true,
	touch_run_id_file::Bool=true,
	verbose=true,
	retry::Int=0,
	dry_run::Bool=false,
	stdout=nothing,
	stderr=nothing,
	stdlog=nothing,
	append::Bool=false
) -> (success::Bool, outputs::Dict{String})

run(p::CmdProgram, inputs, outputs; kwargs...)

run(p::CmdProgram, inputs; kwargs...)
)  # only usable when `p.infer_outputs` is defined, or default outputs are set in `p`.

Run Command Program (CmdProgram) using specified inputs and outputs.

Return (success::Bool, outputs::Dict{String})

p::CmdProgram: the command program template.

inputs::Dict{String} and outputs::Dict{String}: p::CmdProgram stores a command template. In the template, replaceable portions are occupied by keywords, and all keywords can be found at p.inputs and p.outputs string vectors. Here, inputs and outputs are Dict(keyword::String => replacement). The replacements do not have a length limit, unless a keyword refers to a filename (length == 1).

If data types of inputs and outputs are not Dict{String}, they will be converted as far as possible. If the conversion fails, program will throw an error.

dir::AbstractString = "": working directory to run the program and store run_id_file.

check_dependencies::Bool = true: check dependencies for p (p.cmd_dependencies).

skip_when_done::Bool = true: Skip running the program and return true if it has been done before (the run_id_file exists and p.validate_outputs(outputs) passes.)

touch_run_id_file::Bool = true: If true, touch a unique run ID file, which indicate the program is successfully run with given inputs and outputs. If false, the next time running the program, skip_when_done=true will not take effect.

verbose = true: If true or :all, print all info and error messages. If :min, print minimum info and error messages. If false or :none, print error messages only.

retry::Int = 0: If failed, retry for INT times.

dry_run::Bool = false: do not run the command, return (command::AbstractCmd, run_id_file::String).

stdout, stderr, stdlog and append: Redirect the program outputs to files. stdlog is the Julia logging of @info, @warn, @error, etc. Caution: If the original command (p.cmd) has redirection, arguments defined here might not be effective for the command.

Note

Redirecting in Julia are not thread safe, so unexpected redirection might be happen if you are running programs in different Tasks or multi-thread mode.

Workflow

Go to the working directory. Establish redirection. (dir, stdout, stderr, stdlog, append).

Validate compatibility between p and inputs/outputs.

Check whether the program has run before. (skip_when_done, p.validate_outputs(outputs))

Check command dependencies. (check_dependencies, p.cmd_dependencies)

Validate inputs. (p.validate_inputs(inputs))

Generate runnable command from p and inputs/outputs. (stdout, stderr, append)

Preparing before running main command. (p.prerequisites(inputs, outputs))

Run command generated in #5.

Validate outputs. (p.validate_outputs(outputs))

Wrap up. (p.wrap_up(inputs, outputs))

Success, touch run id file, and return (success::Bool, outputs::Dict{String}). (touch_run_id_file::Bool)

Example

p = CmdProgram(
	id_file = "id_file",
	inputs = ["input", "input2"],
	outputs = ["output"],
	cmd = `echo input input2 output`
)

inputs = Dict(
	"input" => `in1`,
	"input2" => 2
)

outputs = Dict(
	"output" => "out"
)

run(p, inputs, outputs;
	touch_run_id_file = false
)

run(
	p::JuliaProgram;
	inputs=Dict{String}(),
	outputs=Dict{String}(),
	dir::AbstractString="",
	check_dependencies::Bool=true,
	skip_when_done::Bool=true,
	touch_run_id_file::Bool=true,
	verbose=true,
	retry::Int=0,
	dry_run::Bool=false,
	stdout=nothing,
	stderr=nothing,
	stdlog=nothing,
	append::Bool=false
) -> (success::Bool, outputs::Dict{String})

run(p::JuliaProgram, inputs, outputs; kwargs...)

run(p::JuliaProgram, inputs; kwargs...)
)  # only usable when `p.infer_outputs` is defined, or default outputs are set in `p`.

Run Julia Program (JuliaProgram) using specified inputs and outputs.

Return (success::Bool, outputs::Dict{String})

p::JuliaProgram: the command program template.

inputs::Dict{String} and outputs::Dict{String}: JuliaProgram stores a Julia function, with two arguments inputs::Dict{String}, outputs::Dict{String}. The keys of the two arguments should be the same as p.inputs::Vector{String} and p.outputs::Vector{String}.

If data types of inputs and outputs are not Dict{String}, they will be converted as far as possible. If the conversion fails, program will throw an error. Caution: the returned value of p.main will be assigned to new outputs. Please ensure the returned value of p.main is Dict{String} with proper keys.

dir::AbstractString = "": working directory to run the program and store run_id_file.

check_dependencies::Bool = true: check dependencies for p (p.cmd_dependencies).

skip_when_done::Bool = true: Skip running the program and return true if it has been done before (the run_id_file exists and p.validate_outputs(outputs) passes.)

touch_run_id_file::Bool = true: If true, touch a unique run ID file, which indicate the program is successfully run with given inputs and outputs. If false, the next time running the program, skip_when_done=true will not take effect.

verbose = true: If true or :all, print all info and error messages. If :min, print minimum info and error messages. If false or :none, print error messages only.

retry::Int = 0: If failed, retry for INT times.

dry_run::Bool = false: do not run the command, return (fake_outputs::Dict{String}, run_id_file::String).

stdout, stderr, stdlog and append: Redirect the program outputs to files. stdlog is the Julia logging of @info, @warn, @error, etc. Caution: If the original function (p.main) has redirection, arguments defined here might not be effective for it.

Note

Redirecting in Julia are not thread safe, so unexpected redirection might be happen if you are running programs in different Tasks or multi-thread mode.

Workflow

Go to the working directory. Establish redirection. (dir, stdout, stderr, stdlog, append).

Validate compatibility between p and inputs/outputs.

Check whether the program has run before. (skip_when_done, p.validate_outputs(outputs))

Check command dependencies. (check_dependencies, p.cmd_dependencies)

Validate inputs. (p.validate_inputs(inputs))

Preparing before running main command. (p.prerequisites(inputs, outputs))

Run main function and ***the returned value will be assigned to new outputs***. (outputs = p.main(inputs, outputs))

Validate outputs. (p.validate_outputs(outputs))

Wrap up. (p.wrap_up(inputs, outputs))

Success, touch run id file, and return (success::Bool, outputs::Dict{String}). (touch_run_id_file::Bool)

Example

p = JuliaProgram(
	id_file = "id_file",
	inputs = ["a", "b"],
	outputs = ["out"],
	main = (inputs, outputs) -> begin
		a = inputs["a"]
		b = inputs["b"]
		println("inputs are ", a, " and ", b)
		println("You can also use info in outputs: ", outputs["out"])

		println("The returned value will be assigned to a new outputs")
		return Dict{String,Any}("out" => b^2)
	end
)

inputs = Dict(
	"a" => `in1`,
	"b" => 2
)

outputs = Dict(
	"out" => "will_be_replaced"
)

success, outputs = run(p, inputs, outputs;
	touch_run_id_file = false
) # outputs will be refreshed

Common Methods

Pipelines.infer_outputs — Function

infer_outputs(p::Program, inputs)
infer_outputs(p::Program, inputs, outputs)

Infer the default outputs from p::Program and inputs::Dict{String}.

Pipelines.check_dependency — Method

check_dependency(p::Program)

Check dependencies listed in p.cmd_dependencies.

Command Dependency

Pipelines.CmdDependency — Method

Struct

mutable struct CmdDependency
	exec::Base.Cmd
	test_args::Base.Cmd
	validate_success::Bool
	validate_stdout::Function
	validate_stderr::Function
	exit_when_fail::Bool
end

Methods

CmdDependency(;
	exec::Base.Cmd=``,
	test_args::Base.Cmd=``,
	validate_success::Bool=false,
	validate_stdout::Function=do_nothing,
	validate_stderr::Function=do_nothing,
	exit_when_fail::Bool=true
)

Create Command Dependency (CmdDependency).

Arguments

exec::AbstractCmd: the command to call the dependency.

test_args::AbstractCmd: for testing purposes, the command to be appended to exec.

validate_success::Bool: when checking the dependency, whether to validate the exit code == 0.

validate_stdout::Function: a function takes standard out as String and return the validation result as ::Bool.

validate_stderr::Function: a function takes standard error as String and return the validation result as ::Bool.

exit_when_fail::Bool: if validation fails, whether to throw error and exit.

Example

julia = CmdDependency(
	exec = Base.julia_cmd(),
	test_args = `--version`,
	validate_success = true,
	validate_stdout = x -> occursin(r"^julia version", x),
	validate_stderr = do_nothing,
	exit_when_fail = true
)

check_dependency(julia)

Pipelines.check_dependency — Method

check_dependency(p::CmdDependency) -> Bool

Check CmdDependency by evaluating:

`$(p.exec) $(p.test_args)`

If success, return true.

If fail, return false, or throw DependencyError when p.exit_when_fail set to true.

Pipelines.check_dependency_dir — Method

check_dependency_dir(path::Union{AbstractString,Cmd}; exit_when_false=true) -> Bool

Checke whether a directory exists. Return ::Bool.

Pipelines.check_dependency_file — Method

check_dependency_file(path::Union{AbstractString,Cmd}; exit_when_false=true) -> Bool

Checke whether a file exists. Return ::Bool.

Missing docstring.

Missing docstring for exec. Check Documenter's build log for details.

Utils

Pipelines.to_str — Method

to_str(x) -> String
str(x) -> String

Convert x to String.

x::Cmd: remove backticks (return string(x)[2:end-1]).

x::Nothing: return "".

x::Vector: join elements with "_" as delim.

x::Any: return string(x).

Pipelines.to_cmd — Method

to_cmd(x) -> Cmd

Convert x to Cmd.

Exception: when x::Nothing, return nothing::Nothing.

Base.split — Method

split(c::Cmd) = c.exec :: Vector{String}

Return splitted arguments of Cmd.

Pipelines.replaceext — Method

replaceext(path, replacement::AbstractString)

If the last component of a path contains a dot, leave everything before the dot as usual, and everything after the dot is replaced by replacement. Otherwise, replacement will be appended to path.

If replacement is empty, the last dot will be removed.

Pipelines.removeext — Method

removeext(path)

If the last component of a path contains a dot, leave everything before the dot as usual, and everything including and after the dot is discarded.

Redirection

Pipelines.redirect_to_files — Function

redirect_to_files(f::Function, file; mode="a+")
redirect_to_files(f::Function, outfile, errfile; mode="a+")
redirect_to_files(f::Function, outfile, errfile, logfile; mode="a+")

Redirect outputs of function f to file(s).

xxxfile: File path (AbstractString), nothing or ::IO. nothing means no redirect. Files can be the same.

mode: same as open(..., mode).

Caution: If xxxfile is an IO, it won't be closed. Please use close(io) or JobSchedulers.close_in_future(io, jobs) manually!

Note

Redirecting in Julia are not thread safe, so unexpected redirection might be happen if you are running programs in different Tasks or multi-thread mode.

Pipelines.restore_stdout — Function

No documentation found.

Pipelines.restore_stderr is a Function.

# 1 method for generic function "restore_stderr":
[1] restore_stderr() in Pipelines at /home/runner/work/Pipelines.jl/Pipelines.jl/src/redirection.jl:24

Missing docstring.

Missing docstring for restore_stderr. Check Documenter's build log for details.