Pipelines.jl Documentation

Functions

Missing docstring.

Missing docstring for CmdDependency(). Check Documenter's build log for details.

Pipelines.check_dependency — Method

check_dependency(p::CmdDependency) -> Bool

Check CmdDependency by evaluating $$(p.exec) $(p.test_args)$.

If success, return true.

If fail, return false, or throw DependencyError when p.exit_when_fail set to true.

Missing docstring.

Missing docstring for check_dependency_dir(p). Check Documenter's build log for details.

Missing docstring.

Missing docstring for check_dependency_file(p). Check Documenter's build log for details.

Pipelines.CmdProgram — Method

CmdProgram(;
	name::String               = "Unnamed Command Program",
	id_file::String            = "",
	info_before::String        = "auto",
	info_after::String         = "auto",
	cmd_dependencies::Vector{CmdDependency} = Vector{CmdDependency}(),
	inputs::Vector{String}     = Vector{String}(),
	validate_inputs::Function  = do_nothing,  # positional arguments: inputs::Dict{String, ValidInputTypes}
	prerequisites::Function    = do_nothing,  # positional arguments: inputs, outputs::Dict{String, ValidInputTypes}
	cmd::Base.AbstractCmd      = ``,
	infer_outputs::Function    = do_nothing,  # positional arguments: inputs::Dict{String, ValidInputTypes}
	outputs::Vector{String}    = Vector{String}(),
	validate_outputs::Function = do_nothing  # positional arguments: outputs::Dict{String, ValidInputTypes}
) -> CmdProgram

Command program template. To run a CmdProgram, use run(::CmdProgram; kwargs...).

name::String: Program name.

id_file::String: The prefix of run ID file. To prevent from running the program with the same inputs and outputs twice, it will generate a unique run ID file after a successful run.

info_before::String: Print it when starting the program.

info_after::String: Print it when finishing the program.

cmd_dependencies::Vector{CmdDependency}: Any command dependencies used in the program.

inputs and outputs: keywords (Vector{String}) in cmd that can be replaced when envoking run(::CmdProgram, inputs::Dict{String, Union{Cmd, AbstractString, Number, Array{T,1} where T}}, outputs::Dict{String, Union{Cmd, AbstractString, Number, Array{T,1} where T}}). See details below.

CmdProgram stores a command template. In the template, replaceable portions are occupied by keywords, and all keywords are set in inputs::Vector{String} and outputs::Vector{String}.

To run the program with replaced keywords, you need to use run(::CmdProgram; inputs::Dict{String, Union{Cmd, AbstractString, Number, Array{T,1} where T}}, outputs::Dict{String, Union{Cmd, AbstractString, Number, Array{T,1} where T}}). The data type is different.

validate_inputs::Function: A function to validate inputs. It takes one argument Dict{String, Union{Cmd, AbstractString, Number, Array{T,1} where T}} whose keys are the same as inputs. If validation fail, throw error or return false.

prerequisites: A function to run just before the main command. It prepares necessary things, such as creating directories. It takes two arguments Dict{String, Union{Cmd, AbstractString, Number, Array{T,1} where T}} whose keys are the same as inputs and outputs, respectively.

cmd::AbstractCmd: The main command template. In the template, keywords in inputs::Vector{String} and outputs::Vector{String} will be replaced when envoking run(::CmdProgram, inputs::Dict{String, Union{Cmd, AbstractString, Number, Array{T,1} where T}}, outputs::Dict{String, Union{Cmd, AbstractString, Number, Array{T,1} where T}}).

infer_outputs::Function: A function to infer outputs from inputs. It takes one argument Dict{String, Union{Cmd, AbstractString, Number, Array{T,1} where T}} whose keys are the same as inputs.

validate_outputs::Function: A function to validate outputs. It takes one argument Dict{String, Union{Cmd, AbstractString, Number, Array{T,1} where T}} whose keys are the same as outputs. If validation fail, throw error or return false.

Base.run — Method

run(
	p::CmdProgram;
	inputs::Dict{String}=Dict{String, Cmd}(),
	outputs::Dict{String}=Dict{String, Cmd}(),
	skip_when_done::Bool=true,
	check_dependencies::Bool=true,
	stdout=nothing,
	stderr=nothing,
	append::Bool=false,
	verbose::Bool=true,
	touch_run_id_file::Bool=true,
	dry_run::Bool=false
) -> (success::Bool, outputs::Dict{String})

run(
	p::CmdProgram,
	inputs::Dict{String},
	outputs::Dict{String};
	kwargs...
)

run(
	p::CmdProgram,
	inputs::Dict{String},
	kwargs...
)

Run Command Program (CmdProgram) using specified inputs and outputs.

Return (success::Bool, outputs::Dict{String})

p::CmdProgram: the command program template.

inputs::Dict{String} and outputs::Dict{String}: p::CmdProgram stores a command template. In the template, replaceable portions are occupied by keywords, and all keywords can be found at p.inputs and p.outputs string vectors. Here, inputs and outputs are Dict(keyword::String => replacement::Union{Cmd, AbstractString, Number, Array{T,1} where T}). The replacements do not have a length limit, unless a keyword refers to a filename (length == 1).

skip_when_done::Bool = true: Skip running the program and return true if it has been done before (the run_id_file exists and p.validate_outputs(outputs) passes.)

check_dependencies::Bool = true: check dependencies for p (p.cmd_dependencies).

stdout, stderr and append: Redirect the program output to files, the same behavior as pipeline(cmd; stdout=stdout, stderr=stderr, append=append). Caution: use after checking whether the original command has redirection.

verbose::Bool = true: If true, print info and error messages. If false, print error messages only.

touch_run_id_file::Bool = true: If true, touch a unique run ID file, which indicate the program is successfully run with given inputs and outputs. If false, the next time running the program, skip_when_done=true will not take effect.

dry_run::Bool = false: do not run the command, return (command::AbstractCmd, run_id_file::String).

Workflow

Validate compatibility between p and inputs/outputs.

Check whether the program has run before. (skip_when_done, p.validate_outputs(outputs))

Check command dependencies. (check_dependencies, p.cmd_dependencies)

Validate inputs. (p.validate_inputs(inputs))

Generate runnable command from p and inputs/outputs. (stdout, stderr, append)

Preparing before running main command. (p.prerequisites(inputs, outputs))

Run command generated in #5.

Validate outputs. (p.validate_outputs(outputs))

Success, touch runidfile, and return true. (touch_run_id_file::Bool)