API

Command Program

Pipelines.CmdProgram — Method

Struct

mutable struct CmdProgram
	name::String
	id_file::String
	info_before::String
	info_after::String
	cmd_dependencies::Vector{CmdDependency}
	inputs::Vector{String}
	validate_inputs::Function
	prerequisites::Function
	cmd::Base.AbstractCmd
	infer_outputs::Function
	outputs::Vector{String}
	validate_outputs::Function
	wrap_up::Function
end

Methods

CmdProgram(;
	name::String               = "Unnamed Command Program",
	id_file::String            = "",
	info_before::String        = "auto",
	info_after::String         = "auto",
	cmd_dependencies::Vector{CmdDependency} = Vector{CmdDependency}(),
	inputs::Vector{String}     = Vector{String}(),
	validate_inputs::Function  = do_nothing,  # positional arguments: inputs::Dict{String, ValidInputTypes}
	prerequisites::Function    = do_nothing,  # positional arguments: inputs, outputs::Dict{String, ValidInputTypes}
	cmd::Base.AbstractCmd      = ``,
	infer_outputs::Function    = do_nothing,  # positional arguments: inputs::Dict{String, ValidInputTypes}
	outputs::Vector{String}    = Vector{String}(),
	validate_outputs::Function = do_nothing  # positional arguments: outputs::Dict{String, ValidInputTypes},
	wrap_up::Function          = do_nothing  # positional arguments: inputs, outputs::Dict{String, ValidInputTypes}
) -> CmdProgram

Command program template. To run a CmdProgram, use run(::CmdProgram; kwargs...).

Arguments

name::String: Program name.

id_file::String: The prefix of run ID file. To prevent from running the program with the same inputs and outputs twice, it will generate a unique run ID file after a successful run.

info_before::String: Print it when starting the program.

info_after::String: Print it when finishing the program.

cmd_dependencies::Vector{CmdDependency}: Any command dependencies used in the program.

inputs and outputs: keywords (Vector{String}) in cmd that can be replaced when envoking run(::CmdProgram, inputs::Dict{String, ValidInputTypes}, outputs::Dict{String, ValidInputTypes}). See details below.

CmdProgram stores a command template. In the template, replaceable portions are occupied by keywords, and all keywords are set in inputs::Vector{String} and outputs::Vector{String}.

To run the program with replaced keywords, you need to use run(::CmdProgram; inputs::Dict{String, ValidInputTypes}, outputs::Dict{String, ValidInputTypes}). The data type is different.

validate_inputs::Function: A function to validate inputs. It takes one argument Dict{String, ValidInputTypes} whose keys are the same as inputs. If validation fail, throw error or return false.

prerequisites: A function to run just before the main command. It prepares necessary things, such as creating directories. It takes two arguments Dict{String, ValidInputTypes} whose keys are the same as inputs and outputs, respectively.

cmd::AbstractCmd: The main command template. In the template, keywords in inputs::Vector{String} and outputs::Vector{String} will be replaced when envoking run(::CmdProgram, inputs::Dict{String, ValidInputTypes}, outputs::Dict{String, ValidInputTypes}).

infer_outputs::Function: A function to infer outputs from inputs. It takes one argument Dict{String, ValidInputTypes} whose keys are the same as inputs.

validate_outputs::Function: A function to validate outputs. It takes one argument Dict{String, ValidInputTypes} whose keys are the same as outputs. If validation fail, throw error or return false.

wrap_up::Function: the last function to run. It takes two arguments Dict{String, ValidInputTypes} whose keys are the same as inputs and outputs, respectively.

Example

p = CmdProgram(
	cmd_dependencies = [julia],
	id_file = "id_file",
	inputs = ["input", "input2"],
	outputs = ["output"],
	cmd = `echo input input2 output`
)

inputs = Dict(
	"input" => `in1`,
	"input2" => 2
)

outputs = Dict(
	"output" => "out"
)

run(p, inputs, outputs;
	touch_run_id_file = false
)

Base.run — Method

run(
	p::CmdProgram;
	inputs::Dict{String}=Dict{String, Cmd}(),
	outputs::Dict{String}=Dict{String, Cmd}(),
	skip_when_done::Bool=true,
	check_dependencies::Bool=true,
	stdout=nothing,
	stderr=nothing,
	append::Bool=false,
	verbose::Bool=true,
	touch_run_id_file::Bool=true,
	dry_run::Bool=false
) -> (success::Bool, outputs::Dict{String})

run(
	p::CmdProgram,
	inputs::Dict{String},
	outputs::Dict{String};
	kwargs...
)

run(
	p::CmdProgram,
	inputs::Dict{String},
	kwargs...
)  # only usable when `p.infer_outputs` is defined.

Run Command Program (CmdProgram) using specified inputs and outputs.

Return (success::Bool, outputs::Dict{String})

p::CmdProgram: the command program template.

inputs::Dict{String} and outputs::Dict{String}: p::CmdProgram stores a command template. In the template, replaceable portions are occupied by keywords, and all keywords can be found at p.inputs and p.outputs string vectors. Here, inputs and outputs are Dict(keyword::String => replacement::Union{Cmd, AbstractString, Number, Array{T,1} where T}). The replacements do not have a length limit, unless a keyword refers to a filename (length == 1).

skip_when_done::Bool = true: Skip running the program and return true if it has been done before (the run_id_file exists and p.validate_outputs(outputs) passes.)

check_dependencies::Bool = true: check dependencies for p (p.cmd_dependencies).

stdout, stderr and append: Redirect the program output to files, the same behavior as pipeline(cmd; stdout=stdout, stderr=stderr, append=append). Caution: use after checking whether the original command has redirection.

verbose::Bool = true: If true, print info and error messages. If false, print error messages only.

touch_run_id_file::Bool = true: If true, touch a unique run ID file, which indicate the program is successfully run with given inputs and outputs. If false, the next time running the program, skip_when_done=true will not take effect.

dry_run::Bool = false: do not run the command, return (command::AbstractCmd, run_id_file::String).

Workflow

Validate compatibility between p and inputs/outputs.

Check whether the program has run before. (skip_when_done, p.validate_outputs(outputs))

Check command dependencies. (check_dependencies, p.cmd_dependencies)

Validate inputs. (p.validate_inputs(inputs))

Generate runnable command from p and inputs/outputs. (stdout, stderr, append)

Preparing before running main command. (p.prerequisites(inputs, outputs))

Run command generated in #5.

Validate outputs. (p.validate_outputs(outputs))

Success, touch run id file, and return true. (touch_run_id_file::Bool)

Example

p = CmdProgram(
	cmd_dependencies = [julia],
	id_file = "id_file",
	inputs = ["input", "input2"],
	outputs = ["output"],
	cmd = `echo input input2 output`
)

inputs = Dict(
	"input" => `in1`,
	"input2" => 2
)

outputs = Dict(
	"output" => "out"
)

run(p, inputs, outputs;
	touch_run_id_file = false
)

Pipelines.infer_outputs — Method

infer_outputs(p::CmdProgram, inputs::Dict{String})

Infer the default outputs from p::CmdProgram and inputs::Dict{String}.

Command Dependency

Pipelines.CmdDependency — Method

Struct

mutable struct CmdDependency
	exec::Base.AbstractCmd
	test_args::Base.AbstractCmd
	validate_success::Bool
	validate_stdout::Function
	validate_stderr::Function
	exit_when_fail::Bool
end

Methods

CmdDependency(;
	exec::Base.AbstractCmd=``,
	test_args::Base.AbstractCmd=``,
	validate_success::Bool=false,
	validate_stdout::Function=do_nothing,
	validate_stderr::Function=do_nothing,
	exit_when_fail::Bool=true
)

Create Command Dependency (CmdDependency).

Arguments

exec::AbstractCmd: the command to call the dependency.

test_args::AbstractCmd: for testing purposes, the command to be appended to exec.

validate_success::Bool: when checking the dependency, whether to validate the exit code == 0.

validate_stdout::Function: a function takes standard out as String and return the validation result as ::Bool.

validate_stderr::Function: a function takes standard error as String and return the validation result as ::Bool.

exit_when_fail::Bool: if validation fails, whether to throw error and exit.

Example

julia = CmdDependency(
	exec = Base.julia_cmd(),
	test_args = `--version`,
	validate_success = true,
	validate_stdout = x -> occursin(r"^julia version", x),
	validate_stderr = do_nothing,
	exit_when_fail = true
)

check_dependency(julia)

Pipelines.check_dependency — Method

check_dependency(p::CmdDependency) -> Bool

Check CmdDependency by evaluating $$(p.exec) $(p.test_args)$.

If success, return true.

If fail, return false, or throw DependencyError when p.exit_when_fail set to true.

Pipelines.check_dependency_dir — Method

check_dependency_dir(path::Union{AbstractString,Cmd}; exit_when_false=true) -> Bool

Checke whether a directory exists. Return ::Bool.

Pipelines.check_dependency_file — Method

check_dependency_file(path::Union{AbstractString,Cmd}; exit_when_false=true) -> Bool

Checke whether a file exists. Return ::Bool.

Utils

Pipelines.to_str — Method

to_str(x) -> String
str(x) -> String

Convert x to String.

x::Cmd: remove backticks (return string(x)[2:end-1]).

x::Nothing: return "".

x::Vector: join elements with "_" as delim.

x::Any: return string(x).

Pipelines.to_cmd — Method

to_cmd(x) -> Cmd

Convert x to Cmd.

Exception: when x::Nothing, return nothing::Nothing.

Base.split — Method

split(c::Cmd) = c.exec :: Vector{String}

Return splitted arguments of Cmd.

Pipelines.replaceext — Method

replaceext(path, replacement::AbstractString)

If the last component of a path contains a dot, leave everything before the dot as usual, and everything after the dot is replaced by replacement. Otherwise, replacement will be appended to path.

If replacement is empty, the last dot will be removed.

Pipelines.removeext — Method

removeext(path)

If the last component of a path contains a dot, leave everything before the dot as usual, and everything including and after the dot is discarded.